Avoid costly `set(group)` operation in finalize_target.
As pointed out in #254, taking a set difference happens to be an expensive operation for large datasets. Specifically, taking `set(group)` lists all files in the bucket, which is currently implemented inefficiently.

In this patch, we find the set intersection with a filter operation. This should be faster since it only checks if a smaller subset of files exists.