Prevent loss of metadata by making the `finalize_target` step atomic.
In debugging why metadata was missing from converting datasets (e.g. #318), I discovered how data may be lost in the `finalize_target` step. Specifically, in the `consolidate_dimension_coordinates` block, if there is an error in the `rmdir` operation during the `group.array` overwrite (as I've consistently experienced), the new array's attributes will be lost (it gets written in a subsequent step). Instead of crashing the pipeline, on the subsequent runs of this step, the data is copied over successfully without the array attributes, which were never persistent. In a nutshell, the `finalize_target` step performs writes in a non-idempotent way, leading to subtle errors that cause attributes to be dropped on the final Zarr. (This, at least, is my leading hypothesis).

To address this issue, I've made an attempt to make the `consolidate_dimension_coordinates` block of `finalize_target` atomic. Per @shoyer's advice, we now first write the data to a new Zarr array, copy the metadata, then rename it to the old dimension. Currently, this approach is a work-in-progress. I will validate the fix end-to-end on a dataset, and then gradually work towards writing unit tests.